+++ src/hooks/usePendingPress.ts
@@ -1,9 +1,4 @@
 import * as React from 'react'
-import {
-  type SharedValue,
-  useSharedValue,
-  withTiming
-} from 'react-native-reanimated'
 
 import { showError } from '../components/services/AirshipInstance'
 import { triggerHaptic } from '../util/haptic'
@@ -18,7 +13,7 @@ export type OnPress = () => void | Promise<void>
 export function usePendingPress(onPress?: OnPress): [boolean, () => void] {
   const [pending, setPending] = React.useState(false)
 
-  function handlePress() {
+  function handlePress(): void {
     if (onPress == null || pending) return
     triggerHaptic('impactLight')
 
@@ -29,7 +24,7 @@ export function usePendingPress(onPress?: OnPress): [boolean, () => void] {
         () => {
           setPending(false)
         },
-        error => {
+        (error: unknown) => {
           setPending(false)
           showError(error)
         }
@@ -38,38 +33,4 @@ export function usePendingPress(onPress?: OnPress): [boolean, () => void] {
   }
 
   return [pending, handlePress]
-}
-
-/**
- * The same as `usePendingPress`, but returns an animated value.
- * This looks nicer, and avoids a re-render.
- */
-export function usePendingPressAnimation(
-  onPress?: OnPress
-): [SharedValue<number>, () => void] {
-  const pending = React.useRef(false)
-  const animation = useSharedValue(0)
-
-  function handlePress() {
-    if (onPress == null || pending.current) return
-
-    const out = onPress()
-    if (out != null && typeof out.then === 'function') {
-      pending.current = true
-      animation.value = withTiming(1)
-      out.then(
-        () => {
-          pending.current = false
-          animation.value = withTiming(0)
-        },
-        error => {
-          pending.current = false
-          animation.value = withTiming(0)
-          showError(error)
-        }
-      )
-    }
-  }
-
-  return [animation, handlePress]
 }